refactor: 불필요한 주석 삭제, 클래스 이동
