+++ src/main/java/roomescape/controller/user/MemberController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.common;
+package roomescape.controller.user;
 
 import java.util.List;
 import org.springframework.http.HttpStatus;
