+++ src/main/java/roomescape/service/MemberService.java
@@ -28,7 +28,6 @@ public class MemberService {
             Member realMemberINDb = memberDao.findById(member.getId());
             return new MemberNameResponse(realMemberINDb);
         } catch (MemberNotFoundException exception) {
-            // TODO DAO에서 안잡고 여기서 잡은 이유 -> dao에서 잡기엔 너무 에러메세지가 authroization로직에서만 유효한 잡기이다.
             throw new InvalidAccessTokenException();
         }
     }
